api/requirements.txt
fastapi
uvicorn
httpx
motor
pymongo
python-dotenv
